requirements.txt
python-dotenv
PyGithub
python-gitlab
GitPython
rich
typer
